requirements.txt
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
cssselect==1.2.0
Django==4.2
django-auto-logout==0.5.1
django-cors-headers==4.1.0
django-csp==3.7
django-debug-toolbar==4.1.0
django-decouple==2.1
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-filter==23.2
django-nine==0.2.7
django-rest-swagger==2.2.0
django-robots==5.0
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-yasg==1.21.5
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
exceptiongroup==1.1.1
fake-useragent==1.1.3
Faker==18.10.1
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
mccabe==0.7.0
minio
mock==5.0.2
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.1
parse==1.19.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.3
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.20.0
pyee==8.2.2
pyflakes==3.0.1
PyJWT==2.7.0
PyMySQL==1.0.3
pyppeteer==1.0.2
pyquery==2.0.0
pyrsistent==0.19.3
pytest==7.3.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-html==0.10.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
simplejson==3.19.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
w3lib==2.1.1
websockets==10.4
zipp==3.15.0
